requirements.txt
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0